Trait amelio::db::repositories::CourseRepository

pub trait CourseRepository {
    fn list_with_names(&self) -> Result<Vec<CourseWithNames>>;

    fn list_names(&self) -> Result<Vec<(i32, String)>>;

    fn get(&self, id: i32) -> Result<Course>;

    fn create(&self, course: NewCourse) -> Result<()>;

    fn enable(&self, id: i32, enable: bool) -> Result<()>;

    fn update(&self, course: EditCourse) -> Result<()>;
}

Course related functionality.

Required methods

fn list_with_names(&self) -> Result<Vec<CourseWithNames>>

List all courses together with their author and tutor names.

fn list_names(&self) -> Result<Vec<(i32, String)>>

List the names (with IDs) of all users that can be assigned to courses.

fn get(&self, id: i32) -> Result<Course>

Get a single course by ID.

fn create(&self, course: NewCourse) -> Result<()>

Create a new course.

fn enable(&self, id: i32, enable: bool) -> Result<()>

Enable or disable an existing course.

fn update(&self, course: EditCourse) -> Result<()>

Update an existing course.