Trait amelio::email::MailRenderer

pub trait MailRenderer {
    fn invitation(&self, name: &str, code: &str) -> (&str, String);
    fn status_change(
        &self,
        name: &str,
        details: StatusDetails<'_>
    ) -> (&str, String);
    fn new_comment(
        &self,
        name: &str,
        details: CommentDetails<'_>
    ) -> (&str, String);
}

A mail renderer creates the subject and body for emails of different purposes.

Required Methods

fn invitation(&self, name: &str, code: &str) -> (&str, String)

Create the invitation email for account activation.

fn status_change(&self, name: &str, details: StatusDetails<'_>) -> (&str, String)

Create the status change email for whenever a ticket status changes.

fn new_comment(&self, name: &str, details: CommentDetails<'_>) -> (&str, String)

Create the new comment email for whenever someone adds a new comment to a ticket.