Module amelio::fairings

Custom fairings (middlewares) that are not specific to any other component.

Structs

Auth

A fairing that handles authentication and authorization for common routes to save boilerplate code. Without this, several routes must provide an extra route for different authorization levels and unauthenticated users.

Csp

A fairing that injects the Content-Security-Policy header if it’s not already present in the HTTP response.

Constants

ADMIN_AUTH_PATHS

CSP_HEADER_NAME

STUDENT_AUTH_PATHS