Module amelio::hashing

Functionality for hashing passwords.

One-way-hashes allow to turn a plaintext password into a hashed version that is unusable without having the original plain password. It protects stolen passwords from being used because only its hashed version is stored and this version can not be turned back into the original one.

Structs

HasherImpl
Main implementation of Hasher.

Traits

Hasher
A hasher can hash and verify passwords with a one-way-hash.

Functions

new_hasher
Create a new hasher.