Trait amelio::hashing::Hasher

pub trait Hasher {
    pub fn hash(&self, password: &str) -> Result<String>;

    pub fn verify(&self, password: &str, hash: &str) -> Result<bool>;
}

A hasher can hash and verify passwords with a one-way-hash.

Required methods

pub fn hash(&self, password: &str) -> Result<String>

Turn a plain passwords into a hashed one.

pub fn verify(&self, password: &str, hash: &str) -> Result<bool>

Verify a plain password against a previously hashed password.

Loading content...