Trait amelio::language::Translate

pub trait Translate {
    fn german(&self) -> &'static str;
}

The translate trait allows for any implementing object to translate itself or its value into different languages.

Required methods

fn german(&self) -> &'static str

Translate to German language.