Module amelio::models

The base models of the system, that services work on.

Structs

Comment
A full comment with all available details.

CommentWithNames
A comment with its creator name included.

Course
A full course with all available details.

CourseWithNames
A course with its author and tutor names included.

EditCourse
An existing course to be updated.

EditTicket
An existing ticket to be updated.

EditUser
An existing user to be updated.

NewComment
A new comment to be added to the system.

NewCourse
A new course to be added to the system.

NewTicket
A new ticket to be added to the system.

NewUser
A basic new user that is not part of the system yet.

Ticket
A full ticket with all available details.

TicketSearch
Different options to search and filter tickets.

TicketWithNames
A ticket with its course and creator names included.

TicketWithRels
A ticket with the same information as TicketWithNames plus the related medium.

User
A full user with all available details.

Enums

Category
The category allows to group Tickets into specific topics.

Medium
A medium contains additional information to locate content for a Ticket. The specific type depends on the TicketType.

MediumType
The kind of medium that is attached to a Ticket based on the TicketType.

NewMedium
A new medium that belongs to a Ticket that is to be added to the system.

Priority
Different priorities of a Ticket.

Role
Different roles that a user can have.

Status
The current status of a ticket.

TicketType
Different types of Tickets. It also decides what kind of medium is attached to a ticket.

Type Definitions

Id
The identifier type for all models.