Module amelio::roles

User authentication and role management related functionality.

Macros

ranked_user_guard 🔒

Structs

AdminUser

The admin user has the Role::Admin role, has access to all features and especially can manage user accounts.

AuthUser

Any user that is authenticated but not checked to have a specific role.

AuthorUser

The author user has the Role::Author role and has the same capabilities as the tutor. This user’s main role is to solve tickets the tutor couldn’t and update media based on the request.

NoUser

The opposite of an AuthUser, a user that is not authenticated.

StudentUser

The student user has the Role::Student role and has the least amount of rights. A student can create and search for tickets only.

TutorUser

The tutor user has the Role::Tutor role and has access to incoming tickets. This user can accept, reject or close reported tickets. In addition he can forward assigned tickets to the author.