Module amelio::routes

All routes for the server.

Modules

assets

Routes for loading static assets.

auth

Authentication related routes.

courses

Course management related routes.

errors

Catchers for custom error pages.

fairing

Special routes that are used by custom fairings.

tickets

Ticket related routes.

users

User management related routes.

Macros

enum_from_request 🔒

from_request 🔒

Structs

Hour

An integer that represents an hour in the range of 0-23.

Minute

An integer that represents a minute in the range of 0-59.

NonEmptyString

A string that is guaranteed to not be empty when parsed from a request param or form value.

PositiveNum

An integer that is guaranteed to equal or greater than 1 when parsed from a request param or form value.

ServerError

A wrapper around anyhow::Error that will print the error and respond with a Status::InternalServerError.

ValidUrl

An Url that’s guaranteed to be valid.

Statics

static_rocket_route_info_for_faq

Rocket code generated static route info.

static_rocket_route_info_for_faq_user

Rocket code generated static route info.

static_rocket_route_info_for_index

Rocket code generated static route info.

static_rocket_route_info_for_index_user

Rocket code generated static route info.

Functions

faq

FAQ page for non-authenticated users, redirecting to the login page.

faq_user

FAQ page for authenticated users.

index

Index page for non-authenticated users, redirecting to the login page.

index_user

Index page for authenticated users.

rocket_route_fn_faq

Rocket code generated wrapping route function.

rocket_route_fn_faq_user

Rocket code generated wrapping route function.

rocket_route_fn_index

Rocket code generated wrapping route function.

rocket_route_fn_index_user

Rocket code generated wrapping route function.

Type Definitions

PositiveId

An ID that is guaranteed to be equal or greater than 1 when parsed from a request param or form value.

Second 🔒

An integer that represents a second in the range of 0-59.