Module amelio::routes::courses

Course management related routes.

Structs

EditCourse
Form data from the course editing form.

NewCourse
Form data from the course creation form.

Statics

static_rocket_route_info_for_edit
Rocket code generated static route info.

static_rocket_route_info_for_enable
Rocket code generated static route info.

static_rocket_route_info_for_list
Rocket code generated static route info.

static_rocket_route_info_for_new
Rocket code generated static route info.

static_rocket_route_info_for_post_edit
Rocket code generated static route info.

static_rocket_route_info_for_post_new
Rocket code generated static route info.

Functions

edit
Course editing form for administrators.

enable
Enable or disable courses as administrator.

list
Course management page for administrators.

new
Course creation form for administrators.

post_edit
Edit course POST endpoint to handle course editing, only for administrators.

post_new
New course POST endpoint to handle course creation, only for administrators.

rocket_route_fn_edit
Rocket code generated wrapping route function.

rocket_route_fn_enable
Rocket code generated wrapping route function.

rocket_route_fn_list
Rocket code generated wrapping route function.

rocket_route_fn_new
Rocket code generated wrapping route function.

rocket_route_fn_post_edit
Rocket code generated wrapping route function.

rocket_route_fn_post_new
Rocket code generated wrapping route function.