Module amelio::routes::tickets

Ticket related routes.

Structs

EditTicket
Form data for the ticket edit form.

NewComment
Form data for the comment form.

NewTicket
Form data for the ticket creation form.

NewTicketData
Form data from the ticket creation form. It contains all available data for differnet kinds of tickets and should never be used directly in a route. Instead use the NewTicket struct.

SearchOptions
Form data for the ticket search form.

Enums

Medium
Different kinds of media from the ticket creation form.

Statics

static_rocket_route_info_for_change_status
Rocket code generated static route info.

static_rocket_route_info_for_edit
Rocket code generated static route info.

static_rocket_route_info_for_forward
Rocket code generated static route info.

static_rocket_route_info_for_list
Rocket code generated static route info.

static_rocket_route_info_for_new
Rocket code generated static route info.

static_rocket_route_info_for_post_add_comment
Rocket code generated static route info.

static_rocket_route_info_for_post_edit
Rocket code generated static route info.

static_rocket_route_info_for_post_new
Rocket code generated static route info.

static_rocket_route_info_for_search
Rocket code generated static route info.

Functions

change_status
Endpoint to change a ticket's status.

edit
Show the details of a ticket from student perspective.

forward
Endpoint to forward a ticket to its course's author.

list
Ticket listing page for students or higher ranked users.

new
Ticket creation form for students or higher ranked users.

post_add_comment
Endpoint to create new ticket comments.

post_edit
Endpoint to update ticket details.

post_new
New ticket POST endpoint to handle ticket creation, for students or higher ranked users.

rocket_route_fn_change_status
Rocket code generated wrapping route function.

rocket_route_fn_edit
Rocket code generated wrapping route function.

rocket_route_fn_forward
Rocket code generated wrapping route function.

rocket_route_fn_list
Rocket code generated wrapping route function.

rocket_route_fn_new
Rocket code generated wrapping route function.

rocket_route_fn_post_add_comment
Rocket code generated wrapping route function.

rocket_route_fn_post_edit
Rocket code generated wrapping route function.

rocket_route_fn_post_new
Rocket code generated wrapping route function.

rocket_route_fn_search
Rocket code generated wrapping route function.

search
Ticket search page for all registered users.