Module amelio::routes::users

User management related routes.

Structs

Activate

Form data from the user activation form.

EditUser

Form data from the user editing form.

NewUser

Form data from the user creation form.

Statics

static_rocket_route_info_for_activate

Rocket code generated static route info.

static_rocket_route_info_for_edit

Rocket code generated static route info.

static_rocket_route_info_for_enable

Rocket code generated static route info.

static_rocket_route_info_for_list

Rocket code generated static route info.

static_rocket_route_info_for_new

Rocket code generated static route info.

static_rocket_route_info_for_post_activate

Rocket code generated static route info.

static_rocket_route_info_for_post_edit

Rocket code generated static route info.

static_rocket_route_info_for_post_new

Rocket code generated static route info.

Functions

activate

User activation page, only accessible to non-authenticated users.

edit

User editing form for administrators.

enable

Enable or disable users as administrator.

list

User management page for administrators.

new

User creation form for administrators.

post_activate

User activation POST endpoint, only accessible to non-authenticated users.

post_edit

Edit user POST endpoint to handle user editing, only for administrators.

post_new

New user POST endpoint to handle user creation, only for administrators.

rocket_route_fn_activate

Rocket code generated wrapping route function.

rocket_route_fn_edit

Rocket code generated wrapping route function.

rocket_route_fn_enable

Rocket code generated wrapping route function.

rocket_route_fn_list

Rocket code generated wrapping route function.

rocket_route_fn_new

Rocket code generated wrapping route function.

rocket_route_fn_post_activate

Rocket code generated wrapping route function.

rocket_route_fn_post_edit

Rocket code generated wrapping route function.

rocket_route_fn_post_new

Rocket code generated wrapping route function.