Struct amelio::services::TicketServiceImpl

struct TicketServiceImpl<TR, CR, UR, MS, MR> where
    TR: TicketRepository,
    CR: CourseRepository,
    UR: UserRepository,
    MS: MailSender,
    MR: MailRenderer,  {
    ticket_repo: TR,
    course_repo: CR,
    user_repo: UR,
    mail_sender: MS,
    mail_renderer: MR,
}

Main implementation of TicketService.

Fields
ticket_repo: TRcourse_repo: CRuser_repo: URmail_sender: MSmail_renderer: MR
Implementations

impl<TR, CR, UR, MS, MR> TicketServiceImpl<TR, CR, UR, MS, MR> where
TR: TicketRepository,
CR: CourseRepository,
UR: UserRepository,
MS: MailSender,
MR: MailRenderer,

fn map_priority(category: Category) -> Priority

Decide the priority of a ticket based on its category.

fn send_status_update(
&self,
ticket: &Ticket,
creator: User,
old: Status,
new: Status
) -> Result<()>

Send an email about a ticket status change.

fn send_comment_update(
&self,
ticket: &Ticket,
creator: User,
writer: User,
comment: &str
) -> Result<()>

Send an email about a new comment for a ticket.