Trait amelio::services::CourseService

pub trait CourseService {
    fn list(&self) -> Result<Vec<CourseWithNames>>;

    fn list_authors_tutors(
        &self
    ) -> Result<(Vec<(i32, String)>, Vec<(i32, String)>)>;

    fn get(&self, id: i32) -> Result<Course>;

    fn create(
        &self, 
        code: String, 
        title: String, 
        author_id: i32, 
        tutor_id: i32
    ) -> Result<()>;

    fn enable(&self, id: i32, enable: bool) -> Result<()>;

    fn update(
        &self, 
        id: i32, 
        title: String, 
        author_id: i32, 
        tutor_id: i32
    ) -> Result<()>;
}

The course service manages courses of the system, like listing existing ones, enable or disable them or adding new ones.

Required methods

fn list(&self) -> Result<Vec<CourseWithNames>>

List all courses together with their author and tutor names.

fn list_authors_tutors(
&self
) -> Result<(Vec<(i32, String)>, Vec<(i32, String)>)>

List all authors and tutors with ID and name.

fn get(&self, id: i32) -> Result<Course>

Get a single course by its ID.

fn create(
&self,
code: String,
title: String,
author_id: i32,
tutor_id: i32
) -> Result<()>

Create a new course in the system.

fn enable(&self, id: i32, enable: bool) -> Result<()>

Enable or disable a course.

fn update(
&self,
id: i32,
title: String,
author_id: i32,
tutor_id: i32
) -> Result<()>

Update the information of a course.

Loading content...