Trait amelio::services::LoginService

pub trait LoginService {
    pub fn login(&self, cred: &Credentials<'_>) -> Result<i32>;
}

The login service manages the user login. Logout is directly handled in the post_logout route because that logic is part of the framework.

Required methods

pub fn login(&self, cred: &Credentials<'_>) -> Result<i32>

Try to login a user and return its database ID if successful.

Loading content...