Trait amelio::services::TicketService

pub trait TicketService {

    fn list(&self) -> Result<Vec<TicketWithNames>>;
    fn list_created(&self, user_id: i32) -> Result<Vec<TicketWithNames>>;
    fn list_assigned(
        &self,
        user_id: i32,
        role: Role
    ) -> Result<Vec<TicketWithNames>>;
    fn list_course_names(&self) -> Result<Vec<(i32, String)>>;
    fn get(&self, id: i32) -> Result<TicketWithNames>;
    fn get_with_rels(
        &self,
        id: i32,
        user_id: i32,
        role: Role
    ) -> Result<TicketWithRels>;
    fn create(&self, ticket: NewTicket, medium: NewMedium) -> Result<i32>;
    fn add_comment(&self, id: i32, writer_id: i32, message: String) -> Result<()>;
    fn update(&self, id: i32, priority: Priority) -> Result<()>;
    fn forward(&self, id: i32) -> Result<()>;
    fn change_status(&self, id: i32, status: Status) -> Result<()>;
    fn search(
        &self,
        role: Role,
        search: &mut TicketSearch
    ) -> Result<Vec<TicketWithNames>>;
    fn can_open(&self, id: i32, user_id: i32, role: Role) -> Result<bool>;

}

The ticket service manages tickets of the system, like listing existing or adding new ones.

Required Methods

fn list(&self) -> Result<Vec<TicketWithNames>>

List all tickets.

fn list_created(&self, user_id: i32) -> Result<Vec<TicketWithNames>>

List all tickets that were created by the given user.

fn list_assigned(&self, user_id: i32, role: Role) -> Result<Vec<TicketWithNames>>

List all tickets that are currently assigned to the given user.

fn list_course_names(&self) -> Result<Vec<(i32, String)>>

List all courses with ID and name.

fn get(&self, id: i32) -> Result<TicketWithNames>

Get a single ticket by its ID.

fn get_with_rels(
&self,
id: i32,
user_id: i32,
role: Role
) -> Result<TicketWithRels>

Get a single ticket together with all relations. If the opening user is a tutor or author and the ticket is still in Status::Open it will be changed to Status::InProgress.

fn create(&self, ticket: NewTicket, medium: NewMedium) -> Result<i32>

Create a new ticket in the system.

fn add_comment(&self, id: i32, writer_id: i32, message: String) -> Result<()>

Add a new comment to a ticket.

fn update(&self, id: i32, priority: Priority) -> Result<()>

Update the details of a ticket.

fn forward(&self, id: i32) -> Result<()>

Forward a ticket to its course’s author.

fn change_status(&self, id: i32, status: Status) -> Result<()>

Change the current status of the ticket.

fn search(
&self,
role: Role,
search: &mut TicketSearch
) -> Result<Vec<TicketWithNames>>

Search for tickets with different criteria.

fn can_open(&self, id: i32, user_id: i32, role: Role) -> Result<bool>

Check whether the user can open a specific ticket.