Module amelio::templates::filters

Custom filters for askama templates.

Functions

opt_select

Compare two values exactly as select but with the first value being optional. If the first value is None, an empty string is returned.

opt_select_ref

Same as opt_select, but takes other as a reference.

opt_str

Return the string representation of a value if it exists or an empty string if it’s None.

select

Compare two values and return selected if they match, an empty string otherwise. This is helpful in pre-selecting a value in HTML <select> elements.

select_ref

Same as select, but takes other as a reference.

srcset

Convert an image URL into a source set with different DPI scaling and output the src and srcset attributes for an <img> element.

timestamp

Convert a UTC timestamp into German time zone and then print it in a custom format fitting for comments.