Function amelio::templates::filters::srcset

pub fn srcset(base: &str) -> Result<String>

Convert an image URL into a source set with different DPI scaling and output the src and srcset attributes for an <img> element.

The different images are expected to be located next to the original image and to have a suffix in the form @<scaling>x. For example, an input of logo.png will create an output as follows:

src="logo.png" srcset="logo.png, logo@1.5x.png 1.5x, logo@2x.png 2x, ..."