Trait amelio::templates::Icon

pub(in templates) trait Icon {
    pub(in templates) fn icon(
        &self
    ) -> &'static str;
}

The icon trait allows to show an icon representation of the implementing object within a template.

This is, similar to the Color trait, mostly useful for enums that should be shown with different icons. The values are CSS classes and bound tot the used icon font.

Required methods

pub(in templates) fn icon(
&self
) -> &'static str

The icon to be shown.

Loading content...