Crate amelio

Amelio - A university project for the IUBH

Amelio is a group project for the IUBH in Germany. It is a ticket system that helps to report and track errors in study media.

The name Amelio is a short version of ameliorate and is another word for improve.

Modules

config
Configuration functions for use with the rocket::Rocket instance.

db
All database related functionality.

email
Functions for sending emails.

fairings
Custom fairings (middlewares) that are not specific to any other component.

hashing
Functionality for hashing passwords.

models
The base models of the system, that services work on.

roles
User authentication and role management related functionality.

routes
All routes for the server.

services
Services of the application which contain the business logic.

templates
All templates that are used to render the web pages of this service.

Functions

main

rocket
Create a new pre-configured Rocket instance.