Module amelio::db

All database related functionality.

Modules

connection

Connection creation for the database.

embedded_migrations

models

Models that map from database rows to structs and back.

repositories

Abstractions over the database for easy access to the data.

schema

Structs

DbMigrations

Database migrations that will be run when attached to a Rocket instance.

TicketData

Enums

Samples

Constants

COURSES_JSON

TICKETS_JSON

USERS_JSON

Traits

QueryExt

Extensions for query fragments in diesel queries.

Functions

create_admin_user

Create the initial admin user.

create_sample_courses

Create several sample courses for testing purposes.

create_sample_tickets

Create several sample tickets for testing purposes.

create_sample_users

Create several sample users for testing purposes.

created

init

Initialize the database by running any outstanding migrations, creating the initial user if none exist.

set_created