Module amelio::db::models

Models that map from database rows to structs and back.

Structs

CommentEntity

A full comment entity equivalent to the comments table.

CourseEntity

A full course entity equivalent to the courses table.

InitCourseEntity

A special new course that is used during first initialization of the database.

InitTicketEntity

A special new ticket that is used during first initialization of the database.

InitUserEntity

A special new user that is used during first initialization of the database.

MediumInteractiveEntity

A full interactive medium entity equivalent to themedium_interactives table, also representing a new entry that can be added to the system.

MediumQuestionaireEntity

A full questionaire medium entity equivalent to themedium_questionaires table, also representing a new entry that can be added to the system.

MediumRecordingEntity

A full recording medium entity equivalent to themedium_recordings table, also representing a new entry that can be added to the system.

MediumTextEntity

A full text medium entity equivalent to themedium_texts table, also representing a new entry that can be added to the system.

NewCommentEntity

A new comment to be added to the database.

NewCourseEntity

A new course to be added to the database.

NewTicketEntity

A new ticket to be added to the database.

NewUserEntity

A new user to be added to the database.

TicketEntity

A full ticket entity equivalent to the tickets table.

UserEntity

A full user entity equivalent to the users table.

Functions

_impl_insertable_for_initcourseentity

_impl_insertable_for_initticketentity

_impl_insertable_for_inituserentity

_impl_insertable_for_mediuminteractiveentity

_impl_insertable_for_mediumquestionaireentity

_impl_insertable_for_mediumrecordingentity

_impl_insertable_for_mediumtextentity

_impl_insertable_for_newcommententity

_impl_insertable_for_newcourseentity

_impl_insertable_for_newticketentity

_impl_insertable_for_newuserentity

_impl_queryable_for_commententity

_impl_queryable_for_courseentity

_impl_queryable_for_mediuminteractiveentity

_impl_queryable_for_mediumquestionaireentity

_impl_queryable_for_mediumrecordingentity

_impl_queryable_for_mediumtextentity

_impl_queryable_for_ticketentity

_impl_queryable_for_userentity