Trait amelio::db::repositories::CourseRepository

pub trait CourseRepository {
    pub fn list_with_names(&self) -> Result<Vec<CourseWithNames>>;

    pub fn list_names(&self) -> Result<Vec<(i32, String)>>;

    pub fn get(&self, id: i32) -> Result<Course>;

    pub fn create(&self, course: NewCourse) -> Result<()>;

    pub fn enable(&self, id: i32, enable: bool) -> Result<()>;

    pub fn update(&self, course: EditCourse) -> Result<()>;
}

Course related functionality.

Required methods

pub fn list_with_names(&self) -> Result<Vec<CourseWithNames>>

List all courses together with their author and tutor names.

pub fn list_names(&self) -> Result<Vec<(i32, String)>>

List the names (with IDs) of all users that can be assigned to courses.

pub fn get(&self, id: i32) -> Result<Course>

Get a single course by ID.

pub fn create(&self, course: NewCourse) -> Result<()>

Create a new course.

pub fn enable(&self, id: i32, enable: bool) -> Result<()>

Enable or disable an existing course.

pub fn update(&self, course: EditCourse) -> Result<()>

Update an existing course.

Loading content...