Trait amelio::db::repositories::TicketRepository

pub trait TicketRepository {
    fn list_with_names(&self) -> Result<Vec<TicketWithNames>>;

    fn list_by_creator_id(
        &self, 
        creator_id: i32
    ) -> Result<Vec<TicketWithNames>>;

    fn list_by_assignee_id(
        &self, 
        assignee_id: i32
    ) -> Result<Vec<TicketWithNames>>;

    fn get_with_names(&self, id: i32) -> Result<TicketWithNames>;

    fn get_with_rels(&self, id: i32) -> Result<TicketWithRels>;

    fn create(
        &self, 
        ticket: NewTicket, 
        priority: Priority, 
        medium: NewMedium
    ) -> Result<()>;

    fn add_comment(&self, comment: NewComment) -> Result<()>;

    fn update(&self, ticket: EditTicket) -> Result<()>;

    fn forward(&self, id: i32) -> Result<()>;

    fn get_status(&self, id: i32) -> Result<Status>;

    fn set_status(&self, id: i32, status: Status) -> Result<()>;

    fn search(&self, search: &TicketSearch) -> Result<Vec<TicketWithNames>>;
}

Ticket related functionality.

Required methods

fn list_with_names(&self) -> Result<Vec<TicketWithNames>>

List all tickets together with their course and creator names.

fn list_by_creator_id(&self, creator_id: i32) -> Result<Vec<TicketWithNames>>

List all tickets by their creator ID.

fn list_by_assignee_id(&self, assignee_id: i32) -> Result<Vec<TicketWithNames>>

List all tickets by their assignee ID. The assignee ID is the tutor or author ID of the course that a ticket belongs to.

fn get_with_names(&self, id: i32) -> Result<TicketWithNames>

Get a single ticket with course and creator names.

fn get_with_rels(&self, id: i32) -> Result<TicketWithRels>

Get a single ticket with all related data.

fn create(
&self,
ticket: NewTicket,
priority: Priority,
medium: NewMedium
) -> Result<()>

Create a new ticket.

fn add_comment(&self, comment: NewComment) -> Result<()>

Add a new comment to an existing ticket.

fn update(&self, ticket: EditTicket) -> Result<()>

Update an existing ticket.

fn forward(&self, id: i32) -> Result<()>

Forward a ticket to its course's author.

fn get_status(&self, id: i32) -> Result<Status>

Get the current status of a ticket.

fn set_status(&self, id: i32, status: Status) -> Result<()>

Set the new status of a ticket.

fn search(&self, search: &TicketSearch) -> Result<Vec<TicketWithNames>>

Search for tickets with different criteria.

Loading content...