Module amelio::email

Functions for sending emails.

Structs

CommentDetails

Detail information to create the new comment email.

Mail

All information needed to send a single mail with the MailSender.

MailRendererImpl

Main implementation of MailRenderer.

SmtpSender

Implementation of MailSender that uses a SMTP client to send emails.

StatusDetails

Detail information to create the status change email.

Traits

MailRenderer

A mail renderer creates the subject and body for emails of different purposes.

MailSender

A mail sender allows to send emails.

Functions

new_mail_renderer

Create a new mail renderer.

new_smtp_sender

Create a new mail sender that uses a SMTP client.