Module amelio::services

Services of the application which contain the business logic.

Structs

CourseServiceImpl

Main implementation of CourseService.

Credentials

The credentials that a user needs to authenticate.

LoginServiceImpl

Main implementation of LoginService.

TicketServiceImpl

Main implementation of TicketService.

UserServiceImpl

Main implementation of UserService.

Traits

CourseService

The course service manages courses of the system, like listing existing ones, enable or disable them or adding new ones.

LoginService

The login service manages the user login. Logout is directly handled in the post_logout route because that logic is part of the framework.

TicketService

The ticket service manages tickets of the system, like listing existing or adding new ones.

UserService

The user service manages users of the system, mainly creation and activation and deactivation.

Functions

course_service

Create a new course service.

login_service

Create a new login service.

ticket_service

Create a new ticket service.

user_service

Create a new user service.

Type Definitions

VecAuthorsTutors

A list of authors and tutors with only their ID and name.