Module amelio::templates

All templates that are used to render the web pages of this service.

Modules

filters
Custom filters for askama templates.

Structs

ActivateUser
Template for the user activation page.

Courses
Template for the user list page.

EditCourse
Template for the edit course page.

EditUser
Template for the edit user page.

Error403
Template for the 403 Forbidden error.

Error404
Template for the 404 Not Found error.

Faq
Template for the FAQ page.

Index
Template for the index page.

Login
Template for the login page.

NewCourse
Template for the new course page.

NewTicket
Template for the new ticket page.

NewUser
Template for the new user page.

SearchTickets
Template for the ticket search page.

TicketDetail
Template for the ticket detail page.

Tickets
Template for the ticket list page.

Users
Template for the user list page.

Enums

MessageCode
Different message codes that can be send as flash messages and translated to different languages.

Traits

Color
The color trait allows to tie a specific color to the object that implements it.

Icon
The icon trait allows to show an icon representation of the implementing object within a template.

Translate
The translate trait allows for any implementing object to translate itself or its value into different languages.